Move irrBitCast to separate header
to avoid polluting all include chains
