+++ irr/include/irrTypes.h
@@ -6,8 +6,6 @@
 
 #include <cstdint>
 #include <cassert>
-#include <cstring>
-#include <type_traits>
 
 //! 8 bit unsigned variable.
 typedef uint8_t u8;
@@ -73,18 +71,6 @@ typedef char fschar_t;
 #else
 #define IRR_DOWN_CAST dynamic_cast
 #endif
-
-// TODO in C++ 20 we'll be able to replace this with bit_cast
-template <typename T, typename U>
-T irrBitCast(const U &x)
-{
-	T res;
-	static_assert(sizeof(T) == sizeof(U));
-	static_assert(std::is_trivially_copyable_v<T>);
-	static_assert(std::is_trivially_copyable_v<U>);
-	std::memcpy(&res, &x, sizeof(U));
-	return res;
-}
 
 //! creates four CC codes used in Irrlicht for simple ids
 /** some compilers can create those by directly writing the
