+++ irr/src/COpenGLDriver.cpp
@@ -2,8 +2,9 @@
 // This file is part of the "Irrlicht Engine".
 // For conditions of distribution and use, see copyright notice in irrlicht.h
 
-#include "COpenGLDriver.h"
 #include <cassert>
+
+#include "COpenGLDriver.h"
 #include "CNullDriver.h"
 #include "EHardwareBufferFlags.h"
 #include "HWBuffer.h"
